CLI virkni til að eyða orði úr grunni o.fl.
hingað til hef ég bara notað sqlite IDE forrit til að eyða orðum úr grunni ef þurft hefur (þ.e. ekki nennt að smíða grunninn upp á nýtt, tekur langan tíma)
bætti við virkni í CLI til að gera það útfrá kennistreng
samhliða því bætti ég virkni í CLI til að fletta upp hvaða orð ef einhver eru útleidd frá gefnu orði, þar sem ekki er hægt/gott að eyða orði sem önnur orð eru skilgreind útfrá
svo bætti ég við JSON default config skrá til að eyða sjálfkrafa þeim orðum sem þarf úr backup skrá sem notuð er í github CI, sjá issue #54

+++ lokaord/database/db.py
@@ -19,37 +19,78 @@ class Base(DeclarativeBase):
 	pass
 
 
-def setup_data_directory(folder_name):
-	# TODO: do we need beeter practices here for preventing directory traversal attack?
-	assert('/' not in folder_name)
-	assert('.' not in folder_name)
+def setup_data_directory(folder_name: str):
+	if '/' in folder_name:
+		raise Exception('Bad name provided!')
+	if '.' in folder_name:
+		raise Exception('Bad name provided!')
 	current_file_directory = os.path.dirname(os.path.realpath(__file__))
 	location_directory = os.path.join(current_file_directory, 'disk')
 	data_directory = os.path.realpath(os.path.join(location_directory, folder_name))
-	assert(data_directory.startswith(location_directory))
+	if not data_directory.startswith(location_directory):
+		raise Exception('Unexpected file directory encountered!')
 	if not os.path.exists(data_directory):
 		os.makedirs(data_directory)
 		logman.Logger.info('Created data directory for "%s".' % (folder_name, ))
 
 
-def backup_sqlite_db_file(folder_name):
+def backup_sqlite_db_file(folder_name: str):
+	if '/' in folder_name:
+		raise Exception('Bad name provided!')
+	if '.' in folder_name:
+		raise Exception('Bad name provided!')
 	timestamp_str = datetime.datetime.utcnow().isoformat()[:19]
-	assert('/' not in folder_name)
-	assert('.' not in folder_name)
 	current_file_directory = os.path.dirname(os.path.realpath(__file__))
 	data_directory = os.path.join(current_file_directory, 'disk', folder_name)
 	sqlite_db_file = os.path.join(data_directory, 'db.sqlite')
 	sqlite_db_bak_file = os.path.join(data_directory, 'db_%s.sqlite' % (timestamp_str, ))
-	assert(os.path.exists(sqlite_db_file))
-	assert(not os.path.islink(sqlite_db_file))
-	assert(os.path.isfile(sqlite_db_file))
+	if not os.path.exists(sqlite_db_file):
+		raise Exception('File does not exist!')
+	if os.path.islink(sqlite_db_file):
+		raise Exception('Folder paths to links are not allowed!')
+	if not os.path.isfile(sqlite_db_file):
+		raise Exception('File to backup not found!')
 	shutil.copy(sqlite_db_file, sqlite_db_bak_file)
 	logman.info(f'Backed up current "db.sqlite" file to "db_{timestamp_str}.sqlite".')
 
 
-def delete_sqlite_db_file(folder_name):
-	assert('/' not in folder_name)
-	assert('.' not in folder_name)
+def use_backup_sqlite_db_file(folder_name: str, filename: str, use_force: bool = False):
+	if '/' in folder_name:
+		raise Exception('Bad name provided!')
+	if '.' in folder_name:
+		raise Exception('Bad name provided!')
+	if '/' in filename:
+		raise Exception('Bad name provided!')
+	if '.' in filename:
+		raise Exception('Bad name provided!')
+	filename_sqlite = '%s.sqlite' % (filename, )
+	current_file_directory = os.path.dirname(os.path.realpath(__file__))
+	data_directory = os.path.join(current_file_directory, 'disk', folder_name)
+	sqlite_db_file = os.path.join(data_directory, 'db.sqlite')
+	sqlite_db_file_bak = os.path.join(data_directory, filename_sqlite)
+	if not os.path.exists(sqlite_db_file_bak):
+		raise Exception('Backup file does not exist!')
+	if os.path.islink(sqlite_db_file_bak):
+		raise Exception('Paths to links are not allowed!')
+	if not os.path.isfile(sqlite_db_file_bak):
+		raise Exception('Backup file not found!')
+	if os.path.exists(sqlite_db_file):
+		if os.path.islink(sqlite_db_file):
+			raise Exception('Folder paths to links are not allowed!')
+		if os.path.isfile(sqlite_db_file):
+			if use_force is True:
+				delete_sqlite_db_file(folder_name)
+			else:
+				raise Exception('File "db.sqlite" already exists, delete or move it.')
+	shutil.copy(sqlite_db_file_bak, sqlite_db_file)
+	logman.info(f'Copied backup "{folder_name}" file "{filename_sqlite}" to "db.sqlite"')
+
+
+def delete_sqlite_db_file(folder_name: str):
+	if '/' in folder_name:
+		raise Exception('Bad name provided!')
+	if '.' in folder_name:
+		raise Exception('Bad name provided!')
 	current_file_directory = os.path.dirname(os.path.realpath(__file__))
 	data_directory = os.path.join('disk', folder_name)
 	sqlite_db_file = os.path.join(data_directory, 'db.sqlite')
@@ -59,14 +100,18 @@ def delete_sqlite_db_file(folder_name):
 			sqlite_db_file,
 		))
 	else:
-		assert(not os.path.islink(abs_sqlite_db_file))
-		assert(os.path.isfile(abs_sqlite_db_file))
+		if os.path.islink(abs_sqlite_db_file):
+			raise Exception('Folder paths to links are not allowed!')
+		if not os.path.isfile(abs_sqlite_db_file):
+			raise Exception('File to delete not found!')
 		os.remove(abs_sqlite_db_file)
 
 
-def create_db_uri(db_name):
-	assert('/' not in db_name)
-	assert('.' not in db_name)
+def create_db_uri(db_name: str):
+	if '/' in db_name:
+		raise Exception('Bad name provided!')
+	if '.' in db_name:
+		raise Exception('Bad name provided!')
 	db_uri = ''
 	use_sqlite = True
 	if use_sqlite:
@@ -79,7 +124,7 @@ def session_has_changes():
 	return bool(Session.new) or bool(Session.dirty) or bool(Session.deleted)
 
 
-def setup_connection(db_uri, db_echo=False):
+def setup_connection(db_uri: str, db_echo: bool = False):
 	global Engine, Session, Base
 	Engine = create_engine(db_uri, echo=db_echo)
 	Session = scoped_session(sessionmaker(autocommit=False, autoflush=False, bind=Engine))
